refactored and combined structure into common flow
