+++ src/Content/Professional/Zippy/Zippy.ts
@@ -1,4 +1,4 @@
-import { ContentEntry } from "../../../Common/ContentEntry"
+import { ContentEntry } from "../../../Common/ContentStructure"
 import sway1 from "./sway1.gif"
 import sway2 from "./sway2.png"
 import zippy1 from "./zippy1.gif"
@@ -7,6 +7,6 @@ import zippy2 from "./zippy2.png"
 
 export const Zippy: ContentEntry = {
     heading: "Zippy",
-    body: "My first startup experiance was at Zippy.ai where our mission was to create a robotics system that could enable autonomus last mile delivery. Here I developed robotics mechanisms to iterate on various morphological solutions. This is where I learned to Tig weld.",
+    body: "My first start up experiance was at Zippy.ai where our mission was to create a robotics system that could enable autonomus last mile delivery. Here I developed robotics mechanisms to iterate on various morphological solutions. This is where I learned to Tig weld.",
     images: [sway1, sway2, zippy1, zippy2]
 }
